Update and Delete methods updated

+++ src/main/java/pt/peralta/shareYourDemo/service/PublicationService.java
@@ -3,6 +3,7 @@ package pt.peralta.shareYourDemo.service;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import pt.peralta.shareYourDemo.entity.Publication;
 import pt.peralta.shareYourDemo.entity.PublicationDTO;
@@ -40,14 +41,24 @@ public class PublicationService {
     public Publication update(Long id, PublicationDTO publicationDTO) throws Exception {
         Publication publication = findById(id);
 
+        if(!ownPublication(publication)) throw new SecurityException("Dont have authorization to Update this publication");
+
         publication.setTitle(publicationDTO.title());
         publication.setDescription(publicationDTO.description());
 
         return repository.save(publication);
     }
 
+    private boolean ownPublication(Publication publication) {
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        return user.getLogin().equals(publication.getCreatedBy());
+    }
+
     public Publication delete(Long id) throws Exception {
         Publication publication = findById(id);
+
+        if(!ownPublication(publication)) throw new SecurityException("Dont have delete to Update this publication");
 
         repository.delete(publication);
